Moved code for writing dts files into dts file code
